Moved all NSLogs over to KTLog.
NB: Removed the debugging options from AbstractConnection so you will need to remove any calls to them. The KTLog mechanism is the replacement instead.

Added a runtime debugging panel to be able to set the logging levels for the differnet logging domains. Command-L in the test application brings up this panel.

FTP now works to josephusjitta.nl, works to samkass.com (which is a windows ftp server) and users.ev1.net



git-svn-id: http://opensource.utr-software.com/source/connection/trunk@63 30aebe3c-6c0d-0410-bd16-8e6c7f3d8b55