Fix a couple of nibs that keep having the wrong size
All over, changed the pattern of getting a byte length of a UTF8String from [string length] to strlen(utf8string)

When fetching a passphrase from the keychain, even a zero-length password is OK, as long as it's there.  This stops me from being asked for my passphrase over an over, even though I saved it (as a empty string) into the keychain.

Slightly modified behavior of -[SSHPassphrase passphraseForPublicKey:account:] to return nil, not @"", if user hit "cancel".  The goal is to cancel the connection if you choose cancel....  this will cause -[SFTPConnection mainThreadPassphrase:] to store nil in myKeychainFingerPrint

In -[SFTPConnection negotiateSSH] around line 288, if the myKeychainFingerPrint (the passphrase) is nil, we abort the connection because the user cancelled.  I put in a #warning asking if we ought to have an error callback.


------ BIG PROBLEM ------


* * * * * * * GREG - However, we are stuck in "Gathering Media Assets".  Is it because we cancel on a background thread??  (Main thread is waiting, KTMediaURLLoader thread is waiting, ConnectionThreadManager thread is waiting...)

* * * * * * * GREG - I'm wondering why myKeychainFingerPrint is used for two different purposes?  both for storing the passphrase (as set in -[SFTPConnection mainThreadPassphrase:] ) and the host fingerprint in -[SFTPConnection keychainFingerPrint].  Does the host's fingerprint actually need to be stored in an ivar?




git-svn-id: http://opensource.utr-software.com/source/connection/trunk@257 30aebe3c-6c0d-0410-bd16-8e6c7f3d8b55