A couple of minor fixes.
One is, on a 553 error, to put the command line into the resulting NSError
to help us get a little more context about what the error was all about.
We can probably spread that to more (all?) of the error codes from
parseCommand:  but I wanted to get GREG's opinion on doing that.

Also pointed out a minor error in negotiateSSH  with a #warning --
not sure how to fix.