Minor change across a bunch of files -- just clarifying when +load is used instead of +initialize why we are doing it; moved a few lines from +load when +initialize is a better candidate.
In general, you only want to use +load for automatic registration of a
class somewhere, not general initialization of statics, etc.

This is due to the following advice:
"Due to the amount of uncertainty about the environment at the point that
load is invoked, you should avoid using load whenever possible. All
class-specific initialization should be done in the class's initialize
method."



git-svn-id: http://opensource.utr-software.com/source/connection/trunk@425 30aebe3c-6c0d-0410-bd16-8e6c7f3d8b55