store an NSError in the CKTransferController of whatever fatal error there might have been, so its client can get the info.
Some error-handling flow cleanup of CKTransferController.  Don't send
the finished content generation message if there was an error.

For the most part, I've tried to move a lot of UI control into the
client of CKTransferController, so that you don't have both client and
CKTC updating strings, etc.  I'm having CKTC update the status, but
the client updates the title, the buttons, etc.

Bad password now sets a fatal error and ends in the usual error way.

(Note, however, that at least in FTP, we get the bad password notification
AFTER the successful connection message, so it will have already started
generating content.)

Abstract connection: put cause's error into the resulting error to help
with context if there is an error.

changed prototype of
- (void)recursivelyDeleteDirectory:(NSString *)path deletionNotificationDelegate:(id)sender;

back to 

- (void)recursivelyDeleteDirectory:(NSString *)path;

since the former is not actually implemented.






git-svn-id: http://opensource.utr-software.com/source/connection/trunk@464 30aebe3c-6c0d-0410-bd16-8e6c7f3d8b55