Encapsulating several areas that were clearing out the connections into a method 'forceDisconnectAll'
Fixed problem of window title not showing up until later by waiting
for main thread to perform the selector - no slowdown noticed.

Trying to better deal with the timeout issue when connect method
never received.  I think this problem will go away soon anyhow when
I've cleared out the leak problem

Fix copy warning -- it's copyWithZone that is in the NSCopying protocol!

Posting a bogus event to ensure that the state change is noticed in the
run loop.  Not sure if this is totally needed though...

Better dealing with connection so that it's set as a weak ref if it's
provided by the delegate.  On dealloc, which calls forceDisconnectAll,
the connection will be properly released.



+ tidy up some formatting :-)