Restore retaining of CKInternalTransferRecord --
Greg said:
"Yeah, the problem won't be seen within Sandvox because CKTransferController retains the "delegate" which is at the moment always a CKTransferRecord. If an app only uses the original upload/download methods that don't take a delegate, then internally the delegate will be created which is a CKTR. If the internal transfer record doesn't retain it then you can see that when the transfer starts up, the delegate will be a dangling pointer and will crash. Because the internal transfer record is a private class, convention doesn't have to apply."