upload from data now pays attention to success in writing file, does error callback, etc. following pattern of upload from file.
Progress/Warning cell now properly shows warning triangle in a folder
if any of its children have warnings.

Fixed recordForFullPath:withRoot: to just call the recursive support
method; now it works (it was not comparing the first element before!)


OPEN ISSUES STILL:

When *no* files get uploaded, maybe we should change to be a fatal error!


Not sure about the file manager delegate "handler" for the fileManager:shouldProceedAfterError: callback -- it seems like we're using it in a few cases, and not in others.

Wondering -- sometimes we use myFileManager, sometimes we use [NSFileManager defaultManager] .... on purpose, or unintentional?