Clarify in all uses of CFStreamCreatePairWithSocketToCFHost that the streams are created with the "Create Rule". Taking out extra retains in one case.
Attempting to not write to the output stream until there is space avaialble. 
This is sort of a stopgap attempt; not sure if it will actually work;
ideally we should heed the available events.

Temporarily turning off FTP verification; this might be interfering with
upload stream functions.

Don't do transferControllerDidFinish: twice when there is a fatal error.