Attempt to solve the mysterious problem of a stream not being ready for data.
With some helpful analysis from Tyler Ballance, it appears that the output
stream, though open at that time, does not have space avaiable yet.  Instead
of waiting for the NSStreamEventHasSpaceAvailable event, we were just starting
to write to that stream, but it would block in those cases where it wasn't
ready yet.  Trying to sleep the thread and loop until "hasSpaceAvailable" returned
true didn't work; neither did a run loop.

So what I did was to factor out the code that writes to the stream to a new
methd "sendQueuedOutput".  At the atempt to send data in "sendData:" we check if
the streadm already is open and has space avaialble; if so, call sendQueuedOutput.
(From my tests, it seems that we have usually already gotten that event at the
time we try to write the data, though obviously not ALWAYS since that was the
original problem!)  If it's not ready to be sent yet, that's OK, we've added it
to the buffer, and as soon as we get the NSStreamEventHasSpaceAvailable, we'll
write out a chunk from the buffer.

I'm not convinced that this is going to solve all our problems but hopefully with
some testing by Dan, Terrence, Greg, and Tyler, we can confirm that this is better.

The one functionality which I have left commented out was the NSError that we would
make, "code:StreamErrorTimedOut", if the thread sleep waited for connection
(and, later, waiting for the stream to become available.)  I like the idea of that
timeout, but since we aren't spinning until it's available, a new, better technique
might be to put in NSTimer, kStreamTimeOutValue seconds in the future, that
we cancel and re-schedule every time through.




git-svn-id: http://opensource.utr-software.com/source/connection/trunk@523 30aebe3c-6c0d-0410-bd16-8e6c7f3d8b55