uncommented this line in -finishSetupForDisplay
/// temporarily turn off FTP verification too  just for simplification
if ([[self connection] isKindOfClass:[FTPConnection class]]) myFlags.verifyTransfers = NO;

so that we no longer do a verification pass for FTP

I was finding that, (exactly?) 50% of the time, the verification connection would fail to disconnect
leaving Sandvox in an inconsistent and confusing state. This change makes it so that Sandvox can
FTP again. Whew!

Obviously, this needs further investigation once 1.2 ships.

Does this break anybody? Does any project but Sandvox use CKTransferController?


git-svn-id: http://opensource.utr-software.com/source/connection/trunk@548 30aebe3c-6c0d-0410-bd16-8e6c7f3d8b55