Change via Brian Amerige
hange : _serverSupport.hasSize = NO;
To: _serverSupport.hasSize = YES;

We do this because when creating connections quickly for the exclusive purposes of using them for a download, the download is queued before the connection actually connects. As a result, it hasn't returned the featurelist yet, so when the download is queued, client apps cannot properly report progress. Changing this value to YES by default makes this transparent --- in the event the server does not support size, it will be corrected when the featurelist is sent.


git-svn-id: http://opensource.utr-software.com/source/connection/trunk@577 30aebe3c-6c0d-0410-bd16-8e6c7f3d8b55