New Controller Object: CKDirectoryTreeController
This class manages a composite view similar to what you see in an NSOpenPanel - list view, column view. There are plans to add an icon view in the future. It is not limited to use just in an panel. There is still a lot of things that aren't implemented, but the core browsing is done. The other stuff will be added over time.

There is a new NSBrowser replacement as NSBrowser didn't cut it. The data source api is very similar to that of NSOutlineView in that there is clear separation of model and view. This class should have no direct ties to other stuff in connection kit, apart from the browser cell so it should be able to be used in other apps.