Fixes for Amazon S3 listing of multi-page directory results. From the ConnectionKit list:
"... ensures that multi-page results (denominated by the isTruncated key) do not contain duplicates in the final contents array we return to the delegate.

S3Connection's use of the prefix parameter in its URI allows it to only receive listings (keys) which potentially fit the bill for the directory we're looking for. Because on S3 we cannot simply as for "/brianamerige/subdirectory", for example, we request "/brianamerige?prefix=/brianamerige/subdirectory". If you're unaware of how listing (and the lack of enforced directory hierarchy) works on S3, you can read more at <http://docs.amazonwebservices.com/AmazonS3/2006-03-01/RESTBucketGET.html>."

git-svn-id: http://opensource.utr-software.com/source/connection/trunk@987 30aebe3c-6c0d-0410-bd16-8e6c7f3d8b55