In SFTP, when we connect, no longer hack around putting the PWD command on top of the queue (improperly), use the pushCommandOnCommandQueue method to have it placed above other command.
Because of this change, we can now safely assume, even on a fresh connection, that when we return to the SFTP prompt ("sftp>"), and thus get a finishedCommand from SFTPTServer, we have properly finished a download –– regardless of how large it is. This also means we determine a download's finished state from reported percentage progress, but from a metric that's actually consistent.

This change will be brought over to the trunk as soon as I get the trunk's SFTP to use our new error handling system. I should begin work on that as soon as Flow 1.1 launches.

git-svn-id: http://opensource.utr-software.com/source/connection/branches/LeopardPlusErrorHandling@1050 30aebe3c-6c0d-0410-bd16-8e6c7f3d8b55