Removes unnecessary key from the recursive download dictionary, which we never made use of. Similarly removes commented-out code that I originally wrote but never finished. (Related to downloading symbolically linked directories. If I were to write it again, I'd tackle it differently anyway, so there's no use in keeping it in the trunk.)
More importantly, we're now checking to see if we downloaded an empty directory in a much, much better way. We just find the appropriate record, and see if the content-count is zero. We used to actually check the descriptions for equivalence, which was downright stupid. (I can say that because I wrote the code myself awhile ago.)