Two major changes here:
(A) We now use the NSURLAuthentication auth methods to create the digest authorization headers. No more crazy algorithm in our code. (This is in the trunk -- I'm not sure how it didn't make its way to the branch.)

(B) We now retain the necessary components to generate a digest authorization header right away for each applicable request, instead of getting rejected for being unauthorized on _every_ request.