Major fixes to CKFileConnection:
(1) First, uploading and downloading actually work now. They also give the delegates/transfer records proper progress/transferred updates.
(2) Uploading and downloading is now independently threaded. Because we're working locally, we have to do this since the transfer methods, _threadedUpload and _threadedDownload, would otherwise block the framework thread. The thread is spun off with detachNewThreadSelector:..., so it's immediately stopped when the transfer fails or is complete.
(3) Ditched the "fc" prefix for the invocation selectors, in place of the more standard "_" prefix, indicating a private method.