Rather large change here. Before merging, to the trunk, please make sure you look over these changes.
In effect, this change allows HTTP connections to upload large files.

To properly handle the upload of large files, I've changed how HTTP connections (WebDAV and S3) chunk their data. The big idea is now that we simply don't attempt to keep all of an upload's content in memory. This is because doing so is impossible on very large files, and otherwise is very slow.

What we now do is send the header first, and then the content separately, if there is any. This allows us to avoid any usage of "appendData" for CKDavUploadFileRequests and CKHTTPPutRequests, as appending data which is large in size, is *extremely* slow or not possible.

To facilitate this, I've changed how StreamBasedConnections holds data that is queued for sending. Where we previously appended the data to a buffer (slow!), we now simply add the new (immutable in the case of uploads) data to a queue. We keep track of our last read location, chunking the data by the same metrics we did before. When we reach the end of a chunk of data in the queue, it's removed from the queue (objectAtIndex:0).

Lastly, I should emphasize that while CKHTTPRequest's instance variable, myContent, is NSMutableData, in CKDAVUploadFileRequest and CKHTTPPutRequest, myContent is immutable NSData. This is because NSMutableData, unlike NSData, reads the ENTIRE file into memory to make it mutable, even if you use the mapped-file initializers. Moreover, the usefulness of mutable-data is lost on uploads, since any mutations to a large piece of data (appending or replacing) is VERY slow. To protect this deviation in the Upload subclasses, the mutable methods on CKHTTPRequest have been overriden, and now throw exceptions. Put simply, HTTP upload requests should NEVER be mutated. 

(Note: I am the practicality of making CKHTTPRequest's myContent immutable itself. This would be preferred to the aforementioned deviation and protection.)

git-svn-id: http://opensource.utr-software.com/source/connection/branches/BrianWorkInProgress@1336 30aebe3c-6c0d-0410-bd16-8e6c7f3d8b55