MASSIVE commit right here. You MUST read this if you're using the "BrianWorkInProgress" branch. As of this commit, my branch is taking CK in a new direction -- a slimmer, more focused direction. Here's the skinny:
(A) Any and all UI elements, short of the droplet, have been removed. This includes the open-panel, transfer controller, and all associated support classes. ConnectionKit is a backend framework, and will remain that way in this branch.

(B) CKDotMacConnection has now been entirely removed. I don't think it's necessary, since it's *really* just a mask over a CKWebDAVConnection, mucking about with the paths to pretend they aren't prefixed with /@username@/. I believe this should be done on the client-side, as it is done in Flow. This should simplify maintenance a bit, too.

(C) Authentication has pretty much been gutted, and brought back to simplicity. We're no longer using the verbose and complicated NSURLAuthenticationChallenge architecture. We now follow a very simple authentication philosophy --- we are provided BOTH the username and password initially on all connections, we attempt to connect, and if we fail, we report it to connection:didOpenAtPath:error, and then disconnect. On the client side, the user/pass may change, and then a new connection may be created for it. As you'll see in the diffs, this is not only much easier to maintain, but it *makes sense*. It's gorgeously simple, and not a tick harder on the client side.

In the coming days, weeks, and months, I will gradually be performing a massive cleaning of CK -- to use consistent designs, practices, and conventions, leverage Obj-C 2.0, compile as 64-bit, and support garbage collection. Good (clean) days ahead!

git-svn-id: http://opensource.utr-software.com/source/connection/branches/BrianWorkInProgress@1420 30aebe3c-6c0d-0410-bd16-8e6c7f3d8b55